refactor(Collapse):添加 CollapseHOC 高阶组件，用来降低创建 collapse subComponent 的成本。
